Fixes - Exodus Wallet & Exceptions
Fixes an Exodus Wallet error that displayed an error to the user, this was because the output path was set to the appdata directory and could not be found by the script.

Also checks for an exception if no wallets have been selected and produces a pop up warning to alert the user.